Points

LAS 1.4 supports 11 point formats which are directly represented by Julia structs. They each implement the abstract type LasPoint{N}, which is parametrised by the point format ID N. Each point format thus has a concrete struct of the form LasPointN <: LasPoint{N} that implements the abstract type for that format ID.

LAS.LasPoint0 — Type

ASPRS LAS point data record format 0

LAS.LasPoint1 — Type

ASPRS LAS point data record format 1

LAS.LasPoint2 — Type

ASPRS LAS point data record format 2

LAS.LasPoint3 — Type

ASPRS LAS point data record format 3

LAS.LasPoint4 — Type

ASPRS LAS point data record format 4

LAS.LasPoint5 — Type

ASPRS LAS point data record format 5

LAS.LasPoint6 — Type

ASPRS LAS point data record format 6

LAS.LasPoint7 — Type

ASPRS LAS point data record format 7

LAS.LasPoint8 — Type

ASPRS LAS point data record format 8

LAS.LasPoint9 — Type

ASPRS LAS point data record format 9

LAS.LasPoint10 — Type

ASPRS LAS point data record format 10

You can get the concrete point format struct and the point format ID for a given LasPoint type with the following helper functions:

LAS.get_point_format — Function

Helper function to get the LAS point format associated with a LAS point record

get_point_format(table)

defined at /home/runner/work/LAS.jl/LAS.jl/src/parse_points.jl:19.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:230.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:231.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:232.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:233.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:234.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:235.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:236.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:237.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:238.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:239.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:240.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:252.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/records.jl:153.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/records.jl:183.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/records.jl:214.

get_point_format(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/records.jl:241.

LAS.get_point_format_id — Function

get_point_format_id(_::Type{T<:LasPoint{N}}) -> Any

Get the numeric format ID corresponding to a particular point format

get_point_format_id(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:249.

get_point_format_id(_)

defined at /home/runner/work/LAS.jl/LAS.jl/src/points.jl:250.